Add VisualBasic support
